[[ LC Builder ]] Implement various things in encoding module
Conflicts:
	engine/src/modules.cpp
	libscript/libscript.xcodeproj/project.pbxproj
	libscript/libstdscript-modules.list
	toolchain/lc-compile/lc-compile.xcodeproj/project.pbxproj
	toolchain/lc-compile/src/module-helper.cpp
