+++ engine/src/exec-filters.cpp
@@ -50,6 +50,122 @@ MC_EXEC_DEFINE_EVAL_METHOD(Filters, SHA1Digest, 2)
 
 ////////////////////////////////////////////////////////////////////////////////
 
+#include "zlib.h"
+
+#define GZIP_HEAD_CRC     0x02 /* bit 1 set: header CRC present */
+#define GZIP_EXTRA_FIELD  0x04 /* bit 2 set: extra field present */
+#define GZIP_ORIG_NAME    0x08 /* bit 3 set: original file name present */
+#define GZIP_COMMENT      0x10 /* bit 4 set: file comment present */
+#define GZIP_RESERVED     0xE0
+#define GZIP_HEADER_SIZE 10
+static char_t gzip_header[GZIP_HEADER_SIZE] = { (char_t)0x1f, (char_t)0x8b,
+    Z_DEFLATED, 0, 0, 0, 0, 0, 0, 3 };
+
+bool MCFiltersCompress(MCDataRef p_source, MCDataRef& r_result)
+{
+    const byte_t *t_src_ptr = MCDataGetBytePtr(p_source);
+    uindex_t t_src_len = MCDataGetLength(p_source);
+    
+    uint32_t size = t_src_len + 12 + GZIP_HEADER_SIZE + 8;
+    size += size / 999;  //dest must be "0.1% larger than (source) plus 12 bytes"
+    MCAutoByteArray t_buffer;
+    if (!t_buffer.New(size))
+        return false;
+    
+    memcpy(t_buffer.Bytes(), gzip_header, GZIP_HEADER_SIZE);
+    z_stream zstrm;
+    memset((char *)&zstrm, 0, sizeof(z_stream));
+    zstrm.next_in = (unsigned char *)t_src_ptr;
+    zstrm.avail_in = t_src_len;
+    zstrm.next_out = (unsigned char *)t_buffer.Bytes() + GZIP_HEADER_SIZE;
+    zstrm.avail_out = size - GZIP_HEADER_SIZE - 8;
+    if (deflateInit2(&zstrm, Z_DEFAULT_COMPRESSION, Z_DEFLATED, -MAX_WBITS, 8, 0) != Z_OK
+        || deflate(&zstrm, Z_FINISH) != Z_STREAM_END
+        || deflateEnd(&zstrm) != Z_OK)
+        return MCErrorThrowGeneric(nil);
+    
+    uint32_t osize = zstrm.total_out + GZIP_HEADER_SIZE;
+    uint32_t check = crc32(0L, Z_NULL, 0);
+    check = crc32(check, (unsigned char *)t_src_ptr, t_src_len);
+    MCSwapInt32HostToBig(check);
+    memcpy(t_buffer.Bytes() + osize, &check, 4);
+    check = t_src_len;
+    MCSwapInt32HostToBig(check);
+    memcpy(t_buffer.Bytes() + osize + 4, &check, 4);
+    
+    t_buffer.Shrink(osize + 8);
+    return t_buffer.CreateDataAndRelease(r_result);
+}
+
+bool MCFiltersIsCompressed(MCDataRef p_source)
+{
+    const char_t *sptr = MCDataGetBytePtr(p_source);
+    return (MCDataGetLength(p_source) >= 10) &&
+    (sptr[0] == gzip_header[0]) &&
+    (sptr[1] == gzip_header[1]) &&
+    (sptr[2] == gzip_header[2]) &&
+    ((sptr[3] & GZIP_RESERVED) == 0);
+}
+
+bool MCFiltersDecompress(MCDataRef p_source, MCDataRef& r_result)
+{
+    const char_t *t_src_ptr = MCDataGetBytePtr(p_source);
+    uindex_t t_src_len = MCDataGetLength(p_source);
+    
+    const char_t *sptr = t_src_ptr;
+    
+    uint32_t startindex = 10;
+    if (sptr[3] & GZIP_EXTRA_FIELD)
+    { /* skip the extra field */
+        uint16_t len;
+        memcpy(&len, &sptr[startindex], 2);
+        MCSwapInt16HostToBig(len);
+        startindex += len;
+    }
+    if (sptr[3] & GZIP_ORIG_NAME) /* skip the original file name */
+        while (startindex < t_src_len && sptr[startindex++])
+            ;
+    if (sptr[3] & GZIP_COMMENT)   /* skip the .gz file comment */
+        while (startindex < t_src_len && sptr[startindex++])
+            ;
+    if (sptr[3] & GZIP_HEAD_CRC) /* skip the header crc */
+        startindex += 2;
+    uint32_t size;
+    memcpy(&size, &sptr[t_src_len - 4], 4);
+    MCSwapInt32HostToBig(size);
+    if (size == 0)
+    {
+        r_result = MCValueRetain(kMCEmptyData);
+        return true;
+    }
+    if (t_src_len < (startindex + 8))
+        return MCErrorThrowGeneric(nil);
+    MCAutoByteArray t_buffer;
+    if (!t_buffer.New(size))
+    {
+        // SMR 1935 no need to abort rest of script...
+        //MCabortscript = False;
+        return false;
+    }
+    
+    z_stream zstrm;
+    memset((char *)&zstrm, 0, sizeof(z_stream));
+    zstrm.next_in = (unsigned char *)sptr + startindex;
+    zstrm.avail_in = t_src_len - startindex - 8;
+    zstrm.next_out = (unsigned char *)t_buffer.Bytes();
+    zstrm.avail_out = size;
+    int err;
+    if (inflateInit2(&zstrm, -MAX_WBITS) != Z_OK
+        || (err = inflate(&zstrm, Z_FINISH)) != Z_STREAM_END
+        && err != Z_OK && err != Z_BUF_ERROR // bug on OS X returns this error
+        || inflateEnd(&zstrm) != Z_OK)
+        return MCErrorThrowGeneric(nil);
+    
+    return t_buffer.CreateDataAndRelease(r_result);
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
 bool MCFiltersIsoToMac(MCDataRef p_source, MCDataRef &r_result)
 {
 	MCAutoByteArray t_buffer;
@@ -305,9 +421,6 @@ void MCFiltersEvalUrlDecode(MCExecContext& ctxt, MCStringRef p_source, MCStringR
 #define BINARY_NOCOUNT (UINDEX_MAX - 1)
 #define BINARY_ALL (UINDEX_MAX)
 
-// From exec-strings.cpp
-extern bool MCStringsEvalTextEncoding(MCStringRef p_encoding, MCStringEncoding &r_encoding);
-
 static bool MCU_gettemplate(MCStringRef format, uindex_t &x_offset, unichar_t &cmd, uindex_t &count, MCStringEncoding &r_encoding)
 {
 	cmd = MCStringGetCharAtIndex(format, x_offset++);    
@@ -325,7 +438,7 @@ static bool MCU_gettemplate(MCStringRef format, uindex_t &x_offset, unichar_t &c
         
         x_offset = t_brace_end + 1;
         
-        if (!MCStringsEvalTextEncoding(*t_encoding_string, r_encoding))
+        if (!MCStringEvalTextEncoding(*t_encoding_string, r_encoding))
             // Encoding not recognised
             return false;    
     }
